refactor: shift CopyrightHolders() and LicenseInfo() to clientversion.cpp
